+++ src/main/java/uk/gov/pay/connector/paymentprocessor/service/CardAuthoriseBaseService.java
@@ -1,5 +1,6 @@
 package uk.gov.pay.connector.paymentprocessor.service;
 
+import com.codahale.metrics.MetricRegistry;
 import io.dropwizard.setup.Environment;
 import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
@@ -11,6 +12,7 @@ import uk.gov.pay.connector.charge.service.ChargeService;
 import uk.gov.pay.connector.chargeevent.dao.ChargeEventDao;
 import uk.gov.pay.connector.common.exception.ConflictRuntimeException;
 import uk.gov.pay.connector.common.exception.OperationAlreadyInProgressRuntimeException;
+import uk.gov.pay.connector.gateway.PaymentProvider;
 import uk.gov.pay.connector.gateway.PaymentProviders;
 import uk.gov.pay.connector.gateway.exception.GenericGatewayRuntimeException;
 import uk.gov.pay.connector.gateway.model.AuthorisationDetails;
@@ -18,6 +20,7 @@ import uk.gov.pay.connector.gateway.model.GatewayError;
 import uk.gov.pay.connector.gateway.model.response.BaseAuthoriseResponse;
 import uk.gov.pay.connector.gateway.model.response.BaseAuthoriseResponse.AuthoriseStatus;
 import uk.gov.pay.connector.gateway.model.response.GatewayResponse;
+import uk.gov.pay.connector.paymentprocessor.model.OperationType;
 
 import javax.persistence.OptimisticLockException;
 import java.util.List;
@@ -29,22 +32,30 @@ import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATIO
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_UNEXPECTED_ERROR;
 import static uk.gov.pay.connector.paymentprocessor.service.CardExecutorService.ExecutionStatus;
 
-public abstract class CardAuthoriseBaseService<T extends AuthorisationDetails> extends CardService<BaseAuthoriseResponse> {
+public abstract class CardAuthoriseBaseService<T extends AuthorisationDetails> {
 
     private static final Logger LOG = LoggerFactory.getLogger(CardAuthoriseBaseService.class);
 
     private final CardExecutorService cardExecutorService;
 
     protected final ChargeService chargeService;
+    protected final ChargeDao chargeDao;
+    protected final ChargeEventDao chargeEventDao;
+    private final PaymentProviders providers;
+    protected final Logger logger = LoggerFactory.getLogger(getClass());
+    protected MetricRegistry metricRegistry;
 
     public CardAuthoriseBaseService(ChargeDao chargeDao, ChargeEventDao chargeEventDao,
                                     PaymentProviders providers,
                                     CardExecutorService cardExecutorService,
                                     ChargeService chargeService,
                                     Environment environment) {
-        super(chargeDao, chargeEventDao, providers, environment);
+        this.chargeDao = chargeDao;
+        this.chargeEventDao = chargeEventDao;
+        this.providers = providers;
         this.cardExecutorService = cardExecutorService;
         this.chargeService = chargeService;
+        this.metricRegistry = environment.metrics();
     }
 
     public GatewayResponse<BaseAuthoriseResponse> doAuthorise(String chargeId, T gatewayAuthRequest) {
@@ -96,6 +107,10 @@ public abstract class CardAuthoriseBaseService<T extends AuthorisationDetails> e
                 .orElseGet(() -> gatewayError
                         .map(this::mapError)
                         .orElse(ChargeStatus.AUTHORISATION_ERROR));
+    }
+
+    protected PaymentProvider<BaseAuthoriseResponse, ?> getPaymentProviderFor(ChargeEntity chargeEntity) {
+        return providers.byName(chargeEntity.getPaymentGatewayName());
     }
 
     private ChargeStatus mapError(GatewayError gatewayError) {
